For keepalive: rejectDuringShutdown, closeSockets
Two new features to deal with clients already connected to the server with idle keep alive connections when an error happens. We've closed the server, but they can still sneak those connections in. With rejectDuringShutdown on, we'll return a 503 along with a connection:close header. With closeSockets, GracefulExpress will manually track all sockets not currently servicing a request.