GracefulExpress: Track _all_ sockets
Seems that Chrome, perhaps due to pre-caching, ends up with a socket connected to your server that never hits the gracefulExpress.middleware. This change ensures that every single socket that connects to the server is tracked by gracefulExpress.

Need to think about how best to play nicely with socket.io now that we're tracking every socket.