Fixes to tests for iojs/node 0.12
Move to built-in agent instead of agentkeepalive.

Tests now handle the big change in node.js/cluster - when no workers are available, connection requests are outright refused. In node 0.10.36 and previous, requests would hang around until a worker was able to service them. The tests leveraged this heavily, making requests immediately after causing an error - on node 10.36 this would result in a delay and a response from the next worker. Now it results in an ECONNREFUSED.